+++ themes/ribir_slim/src/lib.rs
@@ -81,7 +81,7 @@ fn typography_theme() -> TypographyTheme {
         font_size,
         letter_space,
         font_face,
-        overflow: TextOverflow::Clip,
+        overflow: TextOverflow::Overflow,
       },
       decoration: TextDecorationStyle {
         decoration: TextDecoration::NONE,